Change f32 to f64. Remove old vector structs that are no longer used. Implement basic neural network.

+++ src/cost_functions.rs
@@ -1,11 +1,11 @@
-pub fn mean_square_error(lhs: &Vec<f32>, rhs: &Vec<f32>) -> f32 {
+pub fn mean_square_error(lhs: &Vec<f64>, rhs: &Vec<f64>) -> f64 {
     let item_count = lhs.len();
     assert_eq!(item_count, rhs.len());
 
     let mse = lhs.iter()
         .zip(rhs.iter())
         .map(|(a, b)| (a - b).powi(2))
-        .sum::<f32>() / (item_count as f32);
+        .sum::<f64>() / (item_count as f64);
 
     mse
 }
@@ -18,11 +18,11 @@ mod tests {
 
     #[test]
     fn mse_test() {
-        let v1 = &vec![7.0f32, 12.0f32, 25.0f32, -2.0f32];
-        let v2 = &vec![3.0f32, 9.0f32, 24.0f32, -1.0f32];
+        let v1 = &vec![7.0f64, 12.0f64, 25.0f64, -2.0f64];
+        let v2 = &vec![3.0f64, 9.0f64, 24.0f64, -1.0f64];
 
         let actual = mean_square_error(v1, v2);
-        let expected = 6.75f32;
+        let expected = 6.75f64;
 
         assert_eq!(actual, expected);
     }
