+++ src/geoalg/f64_math/complex.rs
@@ -2,7 +2,7 @@ use std::{fmt::Display, ops::*};
 
 #[derive(PartialEq)]
 #[derive(Debug)]
-pub struct Complexf(pub f32, pub f32);
+pub struct Complexf(pub f64, pub f64);
 
 impl Display for Complexf {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
