+++ src/geoalg/f64_math/convolution_kernel.rs
@@ -2,6 +2,6 @@ struct ConvolutionKernel {
     i: u32,
     j: u32,
 
-    data: Vec<f32>
+    data: Vec<f64>
 }
