+++ src/nn/activation_functions.rs
@@ -1,16 +1,16 @@
-use std::f32::consts::E;
+use std::f64::consts::E;
 
 pub struct Activation {
-    pub function: fn(&f32) -> f32,
-    pub derivative: fn(&f32) -> f32
+    pub function: fn(&f64) -> f64,
+    pub derivative: fn(&f64) -> f64
 }
 
 pub const RELU: Activation = Activation {
-    function: |x| if *x <= 0f32 { 0f32 } else { *x },
-    derivative: |x| if *x <= 0f32 { 0f32 } else { 1f32 }
+    function: |x| if *x <= 0f64 { 0f64 } else { *x },
+    derivative: |x| if *x <= 0f64 { 0f64 } else { 1f64 }
 };
 
 pub const SIGMOID: Activation = Activation {
-    function: |x| 1f32 / (1f32 - E.powf(-(*x))),
-    derivative: |x| (*x) * (1f32 - (*x))
+    function: |x| 1f64 / (1f64 - E.powf(-(*x))),
+    derivative: |x| (*x) * (1f64 - (*x))
 };
